+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the `dom_query` crate will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Introduced he Selection::select_matcher_iter method, which returns an iterator over all nodes matching the given matcher, without collecting them into a result vector. Useful for read-only operations
+
+
 ### Changed
 - Minor code refactoring.